requirement.txt
pandas
SQLAlchemy
mysql-connector-python
lxml
python-dotenv